requirements.txt
scikit-learn
xgboost
evidently
mlflow
pandas
numpy
matplotlib
streamlit
plotly
apache-airflow
